1. TP_pairs vector now double (before was int); 2. Fixed statistics crash when objfun term is wrong or there is not data like that to compare; 3. Added TukeyBiWeight with normalized option (normalized residuals); 4. Now also writing the weight to the output file.
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSFIT/branches/devGFshell@479 376c6ad0-4da0-11e1-87d0-276bdbf53620

+++ gemsfit2/src-fit/data_manager.cpp
@@ -921,7 +921,7 @@ void Data_Manager::bson_to_Data_Manager(/* FILE *f, */ const char *data, int pos
 
 void Data_Manager::get_distinct_TP( )
 {
-    vector<int> TP[2];
+    vector<double> TP[2];
     unsigned int i, j;
     bool isfound = false, isfound2 = false;
